fix: fixing the validator
