+++ src/Resources/translations/LleHermesBundle.de.yaml
@@ -188,5 +188,5 @@ modal:
 validator:
     twig_syntax:
         error: Die Vorlage hat Syntaxfehler
-    required_translatable_field:
-        error: Dieses Feld ist in der Standardsprache erforderlich.
+    required_field:
+        error: Dieses Feld ist erforderlich.
